collector/requirements.txt
wget
pandas
psycopg2-binary